+++ backend/ApiAgendamento/src/Routes/RoleRoute.cs
@@ -1,5 +1,7 @@
+using System.Security.Claims;
 using Data;
 using DTOs;
+using Helpers;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Models;
@@ -9,28 +11,37 @@ namespace Routes
   public class RoleRoute
   {
     public static async Task<IResult> CreateRole(
-      [FromServices] ConsultorioDbContext db, 
-      [FromBody] CreateRoleDTO dto
+      [FromServices] ConsultorioDbContext db,
+      [FromBody] CreateRoleDTO dto,
+      ClaimsPrincipal user
     )
     {
+      var roleVerify = VerifyRole.IsAdmin(user);
+      if (!roleVerify) return Results.Forbid();
+
+      if (string.IsNullOrWhiteSpace(dto.Role)) return Results.BadRequest("Nome da role é obrigatório");
+
       Role role = new Role { role = dto.Role };
       db.Add(role);
       await db.SaveChangesAsync();
 
-      return Results.Ok("Role cadastrado");
+      return Results.Created($"/roles/{role.id}", role);
     }
 
     public static async Task<IResult> ChangeRoleName(
-      [FromServices] ConsultorioDbContext db, 
-      string role, 
-      int id
+      [FromServices] ConsultorioDbContext db,
+      ClaimsPrincipal user,
+      [FromBody] UpdateRoleDTO dto
     )
     {
-      var Role = await db.roles.FindAsync(id);
+      var roleVerify = VerifyRole.IsAdmin(user);
+      if (!roleVerify) return Results.Forbid();
 
-      if (Role is null) return Results.NotFound("Role não encontrado");
+      var role = await db.roles.FindAsync(dto.Id);
 
-      Role.role = role;
+      if (role is null) return Results.NotFound("Role não encontrado");
+
+      role.role = dto.Role;
 
       await db.SaveChangesAsync();
 
@@ -38,11 +49,17 @@ namespace Routes
     }
 
     public static async Task<IResult> DeleteRole(
-      [FromServices] ConsultorioDbContext db, 
+      [FromServices] ConsultorioDbContext db,
+      ClaimsPrincipal user,
       int id
     )
     {
-      await db.roles.Where(role => role.id == id).ExecuteDeleteAsync();
+      var roleVerify = VerifyRole.IsAdmin(user);
+      if (!roleVerify) return Results.Forbid();
+
+      var linhasAfetadas = await db.roles.Where(role => role.id == id).ExecuteDeleteAsync();
+
+      if (linhasAfetadas == 0) return Results.NotFound("Role não encontrado");
 
       return Results.NoContent();
     }
